pyproject.toml
[project]
name = "jm"
version = "0.1.0"
description = "Personal job/task management TUI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=1.0.0",
    "rich>=14.0.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
jm = "jm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
